Remove obsolete MemoryLeakTests
The MemoryLeakTests used to test freeing of resources when the resources got freed by the
finalizer. Another test checked freeing by the Dispose method. These tests are now obsolete
because of the improved handling of the Window and Document wrappers: The GeckoDomDocument
is now cached so we always get a reference to the same object, and that object gets
disposed when GeckoWebBrowser gets disposed.